Add support for CAS 5
This is the cheapest fix. I had to make the same change in two places, because
django_cyverse_auth reaches into the cas api response, rather than dealing
with the caslib.py interface, which could abstract things like different api
responses.

In CAS 5, `profile['attributes']` is a dict like:
{ 'lastName': 'Osborn', 'firstname': 'Connor' }

In CAS <= 4, this was modelled as:
[ { 'lastName': 'Osborn' }, { 'firstName': 'Connor' } ]